build(log): restore logging
This restores the dependency to SLF4J that was removed in commit f975fff
as this breaks logging in environments where log4j is not available in
the servlet container.

To accomodate environments that do provide log4j as part of the servlet
container, the POM property `slf4j.scope` is introduced that can be set
to `provided` in build profiles targeting those environments.