feat: implement new item class for managing extents
Adds the new item class "Extent" to the registry that allows managing extents
as register items. Also adds the necessary GUI components to create extent
items.

This is the first step towards managaing geographic extents globally as items
as opposed to being managed as parts of several other item classes.

https://github.com/ISO-TC211/iso-geodetic-registry/issues/49