fix: improve handling of weak and strong Thymeleaf binding
Introduces the model attribute `useWeakBinding` to control whether input
elements are bound strongly using `th:field` or weakly by accurately
setting `th:name` based on the property to be bound.

This prevents problems when fragments are dynamically loaded and the underlying
bean can't be strongly bound, e.g. because a collection entry is added that
does not yet exist in the bean.

https://github.com/ISO-TC211/iso-geodetic-registry/issues/51