refactor: change how item identifiers are calculated and assigned
Changes the way `itemIdentifier`s and, for `IdentifiedItems`, the `identifier`
is calculated and assigned.

The previous way did not allow for item classes that do are not derived from
`IdentifiedItem`. With the introduction of `ExtentItem`, assignment of item
identifiers became inconsistent and duplicate identifiers could occur.

The new assignment algorithm is centered around `RE_RegisterItem`'s
`itemIdentifier` which is available in all item classes. In `IdentifiedItem`s,
the `itemIdentifier` is reused as the `identifier`.