feat: add parameter for filtering register item list by item status
Adds support for a parameter `statusFilter` to the
`/register/{register}/{itemClass}/items` endpoint. A comma-separated list of
`RE_ItemStatus` values can now be passed via that parameter and the returned
list will only contain items that have one of the passed status values.

By default, items with the status `VALID`, `SUPERSEDED`, `RETIRED` and
`INVALID` will be displayed. This also fixes the issue that invalidated items
were not show in the list of contained items.

https://github.com/ISO-TC211/iso-geodetic-registry/issues/70